% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_scen_btns.R
\name{add_scn_btns}
\alias{add_scn_btns}
\title{Action buttons for adding and uploading scenarios (UI function)}
\usage{
add_scn_btns(btn_add_id, drpdwn_btn_id, dwnl_btn_id, file_input_id)
}
\arguments{
\item{btn_add_id}{character string, the \code{inputId} of the button controlling the addition of scenarios}

\item{drpdwn_btn_id}{character string, the \code{inputId} of the button
controlling the dropdown panel with widgets for uploading input parameter
values}

\item{dwnl_btn_id}{character string, the \code{inputId} of the download file widget}

\item{file_input_id}{character string, the \code{inputId} of the upload file widget}
}
\value{
No value returned
}
\description{
Inserts buttons controlling the addition of scenarios and input uploading. It
should be inserted in the \code{title} argument of \code{shinydashboard::tabBox()}
}
