% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_spp_btns.R
\name{add_spp_btns}
\alias{add_spp_btns}
\title{Action buttons for adding and uploading species in windfarms (UI function)}
\usage{
add_spp_btns(
  slctize_id,
  btn_add_id,
  drpdwn_btn_id,
  dwnl_btn_id,
  file_input_id,
  dpdn_close_id
)
}
\arguments{
\item{slctize_id}{character string, the \code{inputId} of the dropdown menu with
the selectize widget for adding species}

\item{dwnl_btn_id}{character string, the \code{inputId} of the download file widget}

\item{file_input_id}{character string, the \code{inputId} of the upload file widget}
}
\value{
No value returned
}
\description{
Inserts buttons controlling the addition of species and input uploading. It
should be inserted in the \code{title} argument of \code{shinydashboard::tabBox()}
}
