% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_wf_btns.R
\name{add_wf_btns}
\alias{add_wf_btns}
\title{Action buttons for adding and uploading Windfarms (UI function)}
\usage{
add_wf_btns(
  dpdn_add_wf_id,
  dpdn_close_id,
  sltz_wf_id,
  dpdn_upld_wf_id,
  dwnl_btn_id,
  file_input_id
)
}
\arguments{
\item{dwnl_btn_id}{character string, the \code{inputId} of the download file widget}

\item{file_input_id}{character string, the \code{inputId} of the upload file widget}

\item{drpdwn_add_wf}{character string, the \code{inputId} of the dropdown menu with wf addition widget}

\item{drpdwn_btn_id}{character string, the \code{inputId} of the button
controlling the dropdown panel with widgets for uploading input parameter
values}
}
\value{
No value returned
}
\description{
Inserts buttons controlling the addition of Windfarms and input uploading. It
should be inserted in the \code{title} argument of \code{shinydashboard::tabBox()}
}
