% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dcmt.R
\docType{data}
\name{FHD_bootstrap_template}
\alias{FHD_bootstrap_template}
\title{Template data for FHD bootstrap replicates}
\format{
data frame, with 500 rows and 201 columns, containing 200 bootstrap
replicates of the distribution of bird flights at 1m height bands, up to
500m above sea level.

\describe{
\item{height_m}{Height above sea level, in metres. First element represents the
0-1 meters height band, and height interval is 1 metre.}
\item{bootId_1}{First bootstrap sample of the proportion of bird flights
within each height interval}
...
\item{bootId_200}{200th bootstrap sample of the proportion of bird flights
within each height interval}
}
}
\usage{
FHD_bootstrap_template
}
\description{
Bootstrap samples of FHD, i.e. the proportion of bird flights at 1 metre
height intervals.
}
\keyword{datasets}
