% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dcmt.R
\docType{data}
\name{FHD_template}
\alias{FHD_template}
\title{Template data for FHD estimates}
\format{
data frame, with 500 rows and 2 columns the distribution of bird
flights at 1m height bands, up to 500m above sea level.
\describe{
\item{height_m}{Height above sea level, in metres. First element represents
the 0-1 meters height band, and height interval is 1 metre.}
\item{prop}{the proportion of bird flights within each height interval}
}
}
\usage{
FHD_template
}
\description{
FHD estimates, i.e. the proportion of birds flights at 1 metre height intervals.
}
\keyword{datasets}
