% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{fPaste}
\alias{fPaste}
\title{Special concatenation}
\source{
\url{https://stackoverflow.com/questions/42456452/changing-vector-to-string-special-separator-for-last-element}
}
\usage{
fPaste(vec)
}
\arguments{
\item{vec}{a character vector}
}
\value{
a character string with elements of \code{vec} concatenated using "," and
"and" separators accordingly
}
\description{
Concatenate multiple strings, using "," to separate intermediate strings and
"and" as the last separator
}
