% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dcmt.R
\docType{data}
\name{monthDens_draws_template}
\alias{monthDens_draws_template}
\title{Template data for monthly density random samples}
\format{
data frame, with 12 columns:
\describe{
\item{January}{Random draws of in-flight bird densities in January.}
...
\item{December}{Random draws of in-flight bird densities in December.}
}
}
\usage{
monthDens_draws_template
}
\description{
Template data for monthly density random samples
}
\keyword{datasets}
