% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dcmt.R
\docType{data}
\name{monthDens_pctls_template}
\alias{monthDens_pctls_template}
\title{Template data for monthly in-flight density reference point estimates}
\format{
data frame, with 13 columns:
\describe{
\item{Percentiles}{numeric values between 0 and 100 (e.g. 25 represents the
25th percentile)}
\item{January}{Point estimates of in-flight bird densities in January.}
...
\item{December}{Point estimates of in-flight bird densities in December}
}
}
\usage{
monthDens_pctls_template
}
\description{
Template data for monthly in-flight density reference point estimates
}
\keyword{datasets}
