% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_plots.R
\name{normDens_ParsPlots}
\alias{normDens_ParsPlots}
\title{------------------------------------------------------------ #
----      Functions for building plot charts            ------
------------------------------------------------------------ #
Normal density plots}
\usage{
normDens_ParsPlots(
  mu,
  stdev,
  fill = "olivedrab",
  xlab,
  refValue_E = mu,
  refValue_SD = stdev
)
}
\description{
------------------------------------------------------------ #
----      Functions for building plot charts            ------
------------------------------------------------------------ #
Normal density plots
}
