% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_remove_tbp.R
\name{remove_tbp}
\alias{remove_tbp}
\title{Remove parent \code{tabPanel} of a clicked button (server function)}
\usage{
remove_tbp(btn_tags, valid_inputs_ids)
}
\arguments{
\item{btn_tags}{character vector, each element a tag composing the button id.
IDs expected to follow the convention \code{btn-'action'-'section'-'section id'}.
E.g. the button for removing the 3rd wind farm scenario expected to be
\code{btn-rmv-wf-3}}

\item{valid_inputs_ids}{character vector with input ids of active buttons}
}
\value{
No returned value
}
\description{
Removes tabpanel of clicked "-" button and selects latest
tabPanel as active. #'
}
