% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dcmt.R
\docType{data}
\name{spp_dflts}
\alias{spp_dflts}
\title{Lookup table with species-specific default parameter values, for the most
common seabird species and a "demo species". Literature references for
default values are also provided.}
\format{
A data-frame with the following columns:
\describe{
\item{spp_id}{The species name/id, to be used in coding}
\item{body_lt}{A list-column of nested data frames with the mean
and sd of body length (m).}
\item{wing_span}{A list-column of nested data frames with the mean
and sd of wing span (m).}
\item{fl_speed}{A list-column of nested data frames with the mean
and sd of flight speed (m/s).}
\item{nct_act}{A list-column of nested data frames with the mean
and sd of nocturnal activity.}
\item{avoid_bsc}{A list-column of nested data frames with the mean
and sd of avoidance probability for the basic models (Option 1).}
\item{avoid_extended}{A list-column of nested data frames with the
mean and sd of avoidance probability for the extended models (Options 2, 3
and 4).}
\item{fl_type}{A list-column of nested data frames with the flight
type.}
\item{upwind_fl}{A list-column of nested data frames with the
proportion of upwind flights.}
\item{prop_crh}{A list-column of nested data frames with the
proportion at collision risk height.}
\item{mth_dens}{A list-column of nested data frames with monthly
mean and sd of in-flight density (birds/km^2) at the windfarm site. Each
data frame consists of 2 rows and 12 columns:
\describe{
\item{January}{the Mean and SD of daytime in-flight bird density in
January, in birds/km^2.}
...
\item{December}{the Mean and SD of daytime in-flight bird density in
December, in birds/km^2.}
}
}
\item{fhd_boot}{A list-column comprising nested data frames with bootstrap
replicates of flight height distribution, i.e. the proportion of bird
flights at 1 metre height intervals. Each data frame consists of 500 rows
and 201 columns, with bootstrap replicates of the distribution of bird
flights at 1m height bands, up to 500m above sea level. Specifically:
\describe{
\item{height_m}{Height above sea level, in metres. First element represents
the 0-1 meters height band, and height interval is 1 metre.}
\item{bootId_1}{First bootstrap sample of the proportion of bird flights
within each height interval}
...
\item{bootId_200}{200th bootstrap sample of the proportion of bird flights
within each height interval}
}
}
\item{fhd_est}{A list-column comprising nested data frames with estimates
of flight height profiles, expressed as the proportion of birds in 1 metre
height intervals. Each data frame contains estimates of the proportion of
flights at 1-metre height intervals above sea level, with columns:
\describe{
\item{height}{Height above sea level, in metres. First element represents
the 0-1 meters height band, and height interval is 1 metre. Highest
height band is currently 499-500 metres.}
\item{est}{Maximum Likelihood estimates of the proportion of flights within
height intervals.}
\item{lcl}{Lower limit of the 95\% CI for the proportion of bird flights
within height intervals.}
\item{ucl}{Upper limit of the 95\% CI for the proportion of bird flights
within height intervals.}
\item{med}{Median of the proportion of bird flights within height intervals.}
}
}
}
}
\source{
Flight Height Distribution data from \href{https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/1365-2664.12191}{Johnston et al (2014)}
}
\usage{
spp_dflts
}
\description{
Lookup table with species-specific default parameter values, for the most
common seabird species and a "demo species". Literature references for
default values are also provided.
}
\keyword{datasets}
