% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dcmt.R
\docType{data}
\name{spp_probdist_pars}
\alias{spp_probdist_pars}
\title{Lookup table listing species features treated as random variables with a
probability distribution. Used to generate UI input widgets dynamically.}
\format{
A data-frame with the following columns:
\describe{
\item{par_name}{used in coding}
\item{par_label}{used for display parameter name in UI}
\item{par_id}{used for inputIDs in e.g. buttons, plots, etc}
\item{par_dist}{probability distribution assumed for the parameter}
}
}
\usage{
spp_probdist_pars
}
\description{
Lookup table listing species features treated as random variables with a
probability distribution. Used to generate UI input widgets dynamically.
}
\keyword{datasets}
