% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dcmt.R
\docType{data}
\name{startup_trb_dwntm}
\alias{startup_trb_dwntm}
\title{Startup values for turbine monthly maintenance downtime}
\format{
A dataframe with 2 rows and 12 columns:
\describe{
\item{January}{the Mean and SD percentage of January whilst turbines were not operating.}
\item{February}{the Mean and SD percentage of February whilst turbines were not operating.}
...
}
}
\usage{
startup_trb_dwntm
}
\description{
Startup values for turbine monthly maintenance downtime
}
\keyword{datasets}
