% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dcmt.R
\docType{data}
\name{startup_wind_avbl}
\alias{startup_wind_avbl}
\title{Startup values for monthly operational wind availability}
\format{
A dataframe with 1 row and 12 columns

\describe{
\item{January}{The percentage of January during which operational wind conditions}
\item{February}{The percentage of February during which wind conditions
allowed for turbine operation}
...
}
\usage{
startup_wind_avbl
}
\description{
Startup values for monthly operational wind availability
}
\keyword{datasets}
