% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_trbn_cfg_inputs.R
\name{trbn_cfg_inputs}
\alias{trbn_cfg_inputs}
\title{Insert inputs for turbine configuration parameters (UI function)}
\usage{
trbn_cfg_inputs(ns, inputs_width = "100\%", is_demo = FALSE)
}
\arguments{
\item{inputs_width}{character string, the width of the numeric input, e.g. '400px',
or '100\%'}
}
\value{
No returned value.
}
\description{
Insert \code{shinydashboard::box} with the following turbine input parameters in
windfarm scenario panel:
\itemize{
\item Nr of blades
\item Rotor radius
\item Air Gap
\item Blade width
}
}
