% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_wf_feat_inputs.R
\name{wf_feat_inputs}
\alias{wf_feat_inputs}
\title{Insert winfarm features inputs}
\usage{
wf_feat_inputs(ns = ns, inputs_width = "100\%")
}
\arguments{
\item{inputs_width}{character string, the width of the numeric input, e.g. '400px',
or '100\%'.}
}
\value{
No returned value.
}
\description{
Insert a \code{shinydashboardPlus::box} in the a windfarm scenario panel with the
following input parameters:
\itemize{
\item Nr of turbines
\item Windfarm latitude
\item Air Gap
\item Blade width
}
}
