% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_remove_tbp.R
\name{remove_tbp}
\alias{remove_tbp}
\title{Remove parent \code{tabPanel} of a clicked button (server function)}
\usage{
remove_tbp(btn_tags, valid_inputs_ids)
}
\arguments{
\item{btn_tags}{character vector, each element a tag composing the button id}

\item{valid_inputs_ids}{character vector with valid ui input ids}
}
\value{
No returned value
}
\description{
Removes tabpanel of clicked "-" button and selects latest
tabPanel as active
}
