% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dcmt.R
\docType{data}
\name{wf_pars_lookup}
\alias{wf_pars_lookup}
\title{Lookup table for windfarm and turbine parameters}
\format{
A dataframe with the followig columns:
\describe{
\item{name}{used in coding}
\item{label}{used for display parameter name in UI}
\item{id}{used for inputIDs in e.g. buttons, plots, etc}
\item{dist}{probability distribution assumed for the parameter}
\item{dflt_mu}{default for the mean}
\item{dflt_sd}{default for the sd}
}
}
\usage{
wf_pars_lookup
}
\description{
Lookup table for windfarm and turbine parameters
}
\keyword{datasets}
